Update HTML/CSS, Added script.js
Updated the web page and the styling for the page. Added script.js back and got rid of the inline JavaScript.
Added an alert window that displays some details for each outage when the outage circle is clicked.
Each outage will be based on a real-world scenario (this will be added later).
Added a command line functionality with five commands (help, clear, escalate, dispatch, troubleshoot).
These commands will be used to deal with the scenarios mentioned above. Depending on what command the user picks they will either successfully deal with the scenario or they will fail.