Update script.js
Updated script.js to add functionality for randomized outages in five cities across the continental United States.
Each outage will display in the alert window with some details.
When an alert bubble is clicked, you will be notified in the alert window of which alert you are currently working on.
Still need to add functionality so that the user can deal with the outage via the command line.