Update script.js
Updated the JavaScript to add some functionality to the application.
Users can now diagnose outages by using the 'diagnostics' command. This command will give more details to the user.
The user can then decide what command to use to deal with the outage.
If the right command is used, the outage bubble will disappear, and the alert window will display a success message.